Try to execute go command in workspace
Unfortunately this won't work as expected because the working directory
of a command `cmd.Dir` can be set to a symlinked directory _but_ the
working directory will immediately resolve to the symlink's target.
Things will work in `bash` and `zsh` as expected because their `pwd` is
a shell builtin but go will use `/bin/pwd` or some kernel function
instead.

At the moment I have no satisfactory answer to this problem.
Recursively copying the project contents into the working space directory will work
but:

- takes some time depending on the project
- doubles the amount of required disk space
- makes it hard to refelect changes back, e.g. `go fmt -w` and so on

:disappointed: